# openapi/throughput-v1.yaml
...
openapi: 3.1.0
info:
  title: Throughput API
  version: 1.0.0
  summary: Multi-tenant B2B sales and operations console — public API.
  description: |
    Read and write access to accounts, contacts, deals, orders, invoices and stock for a
    single workspace.

    ## Authentication

    Send a personal access token as `Authorization: Bearer <token>`. Tokens are issued in
    the application, under **Settings → API tokens**, by an Owner or a Manager
    (FR-API-02), and the plain value is shown exactly once.

    **The workspace is derived from the token, never from the URL** (ADR-008, §18.2). No
    path in this document contains a workspace segment, and there is no header or body
    field that can select one: a token belongs to exactly one workspace for its whole
    life.

    ## Scopes

    Every token carries an explicit list of scopes (FR-API-01). An endpoint whose scope is
    missing answers `403` with the scope named in the message — for example
    `Missing required scope: orders:write`.

    Scopes narrow; they never widen. The role of the member who issued the token
    (§7.4) still applies on top, so a token with `orders:write` issued by a Viewer is
    still refused.

    ## Object identifiers and cross-workspace access

    Identifiers are ULIDs, not sequential integers, so they cannot be enumerated (§18.5).
    A request for an object that belongs to a different workspace answers `404`, never
    `403` — a `403` would confirm that the object exists.

...

...

...

...

...

...

...
servers:
...
security:
  - sanctumToken: []
paths:
...
  /invoices:
    get:
      tags: [Invoices]
      summary: List invoices
      description: Invoices in this workspace, newest first. Requires the `invoices:read` scope.
      operationId: invoices.index
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PerPage"
        - name: status
          in: query
          description: Restrict to one invoice status.
          required: false
          schema:
            type: string
            enum: [draft, sent, paid, overdue, void]
        - name: orderId
          in: query
          description: Restrict to the invoices raised from one order.
          required: false
          schema:
            type: string
      responses:
        "200":
          description: A page of invoices.
          content:
            application/json:
              schema:
                type: object
                required: [data, meta]
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Invoice"
                  meta:
                    $ref: "#/components/schemas/PageMeta"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/MissingScope"
        "429":
          $ref: "#/components/responses/TooManyRequests"
    post:
      tags: [Invoices]
      summary: Raise an invoice from an order
      description: |
        Requires the `invoices:write` scope and an `Idempotency-Key` header (FR-API-03).

        The order must be `confirmed` or `fulfilled` (US-BILL-01) and must not already
        have an invoice that is not void (§12.1). The invoice number is allocated by the
        server; the PDF is rendered asynchronously, so `pdfStatus` starts at `pending`.
      operationId: invoices.store
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInvoiceRequest"
      responses:
        "201":
          description: The invoice that was created.
          headers:
            Idempotent-Replay:
              $ref: "#/components/headers/IdempotentReplay"
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    $ref: "#/components/schemas/Invoice"
        "400":
          $ref: "#/components/responses/IdempotencyKeyMissing"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/MissingScope"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/IdempotencyInProgress"
        "422":
          $ref: "#/components/responses/ValidationFailed"
        "429":
          $ref: "#/components/responses/TooManyRequests"
...
  /inventory:
    get:
      tags: [Stock]
      summary: List stock levels
      description: |
        On-hand, reserved and available quantities per variant and location. Requires the
        `inventory:read` scope.

        This is also the route that tells a consumer which `variantId` and `locationId`
        values exist, which is what `POST /stock-movements` and order lines need.
      operationId: inventory.index
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PerPage"
        - name: variantId
          in: query
          description: Restrict to one product variant.
          required: false
          schema:
            type: string
        - name: locationId
          in: query
          description: Restrict to one stock location.
          required: false
          schema:
            type: string
      responses:
        "200":
          description: A page of stock levels.
          content:
            application/json:
              schema:
                type: object
                required: [data, meta]
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/InventoryLevel"
                  meta:
                    $ref: "#/components/schemas/PageMeta"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/MissingScope"
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /stock-movements:
    get:
      tags: [Stock]
      summary: List stock movements
      description: |
        The append-only stock ledger, newest first (ADR-004). Requires the
        `inventory:read` scope.
      operationId: stock-movements.index
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PerPage"
        - name: variantId
          in: query
          description: Restrict to one product variant.
          required: false
          schema:
            type: string
        - name: locationId
          in: query
          description: Restrict to one stock location.
          required: false
          schema:
            type: string
        - name: reason
          in: query
          description: Restrict to one movement reason.
          required: false
          schema:
            type: string
            enum: [receipt, sale, adjustment, return, transfer]
      responses:
        "200":
          description: A page of stock movements.
          content:
            application/json:
              schema:
                type: object
                required: [data, meta]
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/StockMovement"
                  meta:
                    $ref: "#/components/schemas/PageMeta"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/MissingScope"
        "429":
          $ref: "#/components/responses/TooManyRequests"
    post:
      tags: [Stock]
      summary: Record a stock movement
      description: |
        Requires the `inventory:write` scope and an `Idempotency-Key` header (FR-API-03).

        Only the two human-entered reasons are accepted here: `receipt` and `adjustment`.
        `sale`, `transfer` and `return` are written by the application's own flows and
        carry invariants an external caller cannot satisfy.

        An `adjustment` must carry a note explaining the correction (BR-STOCK-01). A
        movement that would take on-hand below zero is refused. Corrections are new
        movements — the ledger is never updated or deleted.
      operationId: stock-movements.store
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStockMovementRequest"
      responses:
        "201":
          description: The movement that was recorded.
          headers:
            Idempotent-Replay:
              $ref: "#/components/headers/IdempotentReplay"
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    $ref: "#/components/schemas/StockMovement"
        "400":
          $ref: "#/components/responses/IdempotencyKeyMissing"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/MissingScope"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/IdempotencyInProgress"
        "422":
          $ref: "#/components/responses/ValidationFailed"
        "429":
          $ref: "#/components/responses/TooManyRequests"
components:
  securitySchemes:
    sanctumToken:
      type: http
      scheme: bearer
      description: |
        Personal access token issued by Laravel Sanctum, created under Settings → API
        tokens. The token carries the workspace it belongs to and the scopes chosen when
        it was issued (§18.1). The plain value is shown once, at creation, and is not
        recoverable afterwards.
  parameters:
    Page:
      name: page
      in: query
      description: 1-based page number.
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPage:
      name: perPage
      in: query
      description: Rows per page, at most 100.
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 25
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      description: |
        A unique key chosen by the caller, at most 255 characters — a UUID is the usual
        choice. Replaying the same key with the same body returns the first response
        instead of repeating the effect (§18.4).
      required: true
      schema:
        type: string
        maxLength: 255
...
  headers:
...
    RetryAfter:
      description: Seconds to wait before retrying.
      schema:
        type: integer
    IdempotentReplay:
      description: Present and `true` when the body is a replay of an earlier request with the same key.
      schema:
        type: string
  responses:
    Unauthorized:
      description: No token, or a token that is invalid, expired or revoked.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    MissingScope:
      description: |
        One of three things. The token lacks the scope this endpoint requires — the
        message names it and `requiredScope` carries it. Or the member who issued the
        token lacks the permission for this action (§7.4); scopes narrow, they never
        widen. Or the workspace subscription is unpaid or cancelled, which makes the API
        read-only or closed for every role, exactly as it does in the application (§12.2).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ScopeError"
    NotFound:
      description: |
        No such object in this workspace. An object that exists in a different workspace
        answers exactly this, not `403` — the API does not confirm that it exists (§18.5).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    ValidationFailed:
      description: |
        The body failed validation, or the `Idempotency-Key` was reused with a different
        payload (§18.4).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ValidationError"
    IdempotencyKeyMissing:
      description: The `Idempotency-Key` header is required on this endpoint and was absent or too long.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    IdempotencyInProgress:
      description: An earlier request with this `Idempotency-Key` is still running.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    TooManyRequests:
      description: The token exceeded 300 requests in a minute (FR-API-05).
      headers:
        Retry-After:
          $ref: "#/components/headers/RetryAfter"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    Error:
      type: object
      required: [message]
      properties:
        message:
          type: string
          description: A sentence a developer can act on, never a generic failure.
    ScopeError:
      type: object
      required: [message]
      properties:
        message:
          type: string
          examples: ["Missing required scope: orders:write"]
        requiredScope:
          type: string
          description: The scope the token would need.
          examples: ["orders:write"]
    ValidationError:
      type: object
      required: [message]
      properties:
        message:
          type: string
        errors:
          type: object
          description: Field name to the list of problems with it.
          additionalProperties:
            type: array
            items:
              type: string
    PageMeta:
      type: object
      required: [page, perPage, total, lastPage]
      properties:
        page:
          type: integer
        perPage:
          type: integer
        total:
          type: integer
        lastPage:
          type: integer
...
    Invoice:
      type: object
      required: [id, status]
      properties:
        id:
          type: string
          description: ULID.
        invoiceNumber:
          type: [string, "null"]
        orderId:
          type: string
        status:
          type: string
          enum: [draft, sent, paid, overdue, void]
        issueDate:
          type: [string, "null"]
          format: date
        dueDate:
          type: [string, "null"]
          format: date
        currency:
          type: string
          minLength: 3
          maxLength: 3
        subtotal:
          type: number
        taxTotal:
          type: number
        total:
          type: number
        amountPaid:
          type: number
        balanceDue:
          type: number
        pdfStatus:
          type: string
          enum: [pending, ready, failed]
          description: The PDF is rendered in a queue (ADR-013); it is downloadable once this is `ready`.
        voidReason:
          type: [string, "null"]
        voidedAt:
          type: [string, "null"]
          format: date-time
        createdAt:
          type: [string, "null"]
          format: date-time
        updatedAt:
          type: [string, "null"]
          format: date-time
    CreateInvoiceRequest:
      type: object
      required: [order_id]
      properties:
        order_id:
          type: string
          description: An order of this workspace, in status `confirmed` or `fulfilled`.
    InventoryLevel:
      type: object
      required: [variantId, locationId, onHand, reserved, available]
      properties:
        variantId:
          type: string
        sku:
          type: [string, "null"]
        locationId:
          type: string
        locationName:
          type: [string, "null"]
        onHand:
          type: integer
        reserved:
          type: integer
          description: Held for confirmed orders that are not yet shipped.
        available:
          type: integer
          description: On-hand minus reserved. Never negative by design (BR-STOCK-04).
        updatedAt:
          type: [string, "null"]
          format: date-time
    StockMovement:
      type: object
      required: [id, delta, reason]
      properties:
        id:
          type: string
          description: ULID.
        variantId:
          type: string
        locationId:
          type: string
        delta:
          type: integer
          description: Signed. Never zero.
        reason:
          type: string
          enum: [receipt, sale, adjustment, return, transfer]
        refType:
          type: [string, "null"]
          description: What caused the movement, when it came from another record.
        refId:
          type: [string, "null"]
        note:
          type: [string, "null"]
        createdAt:
          type: [string, "null"]
          format: date-time
    CreateStockMovementRequest:
      type: object
      required: [variant_id, location_id, delta, reason]
      properties:
        variant_id:
          type: string
        location_id:
          type: string
        delta:
          type: integer
          description: Signed, non-zero.
        reason:
          type: string
          enum: [receipt, adjustment]
          description: Only human-entered reasons are accepted over the API.
        note:
          type: [string, "null"]
          maxLength: 500
          description: Required when `reason` is `adjustment` (BR-STOCK-01).